@@ -34,15 +34,14 @@
   "homepage": "https://github.com/kuuurt13/ng-block-ui#readme",
   "dependencies": {},
   "devDependencies": {
-    "@angular/common": "6.0.7",
-    "@angular/compiler": "6.0.7",
-    "@angular/compiler-cli": "6.0.7",
-    "@angular/core": "6.0.7",
-    "@angular/forms": "6.0.7",
-    "@angular/http": "6.0.7",
-    "@angular/platform-browser": "6.0.7",
-    "@angular/platform-browser-dynamic": "6.0.7",
-    "@angular/router": "6.0.7",
+    "@angular/common": "^9.1.9",
+    "@angular/compiler": "^9.1.9",
+    "@angular/compiler-cli": "^9.1.9",
+    "@angular/core": "^9.1.9",
+    "@angular/forms": "^9.1.9",
+    "@angular/platform-browser": "^9.1.9",
+    "@angular/platform-browser-dynamic": "^9.1.9",
+    "@angular/router": "^9.1.9",
     "@types/jasmine": "^2.8.9",
     "babel-polyfill": "^6.23.0",
     "es6-shim": "^0.35.1",
@@ -55,19 +54,20 @@
     "karma-jasmine": "^1.0.2",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^1.7.0",
-    "ng-packagr": "^4.7.1",
+    "ng-packagr": "^9.1.5",
     "raw-loader": "^0.5.1",
     "reflect-metadata": "^0.1.3",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.5.3",
     "ts-loader": "^0.8.2",
+    "tslib": "^1.10.0",
     "tslint": "^3.13.0",
     "tslint-loader": "^2.1.5",
-    "typescript": "2.7.2",
+    "typescript": "^3.6.5",
     "typings": "^1.3.1",
     "watchman": "^0.1.8",
     "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
-    "zone.js": "^0.8.26"
+    "webpack-dev-server": "^1.16.5",
+    "zone.js": "^0.10.3"
   },
   "$schema": "./node_modules/ng-packagr/package.schema.json",
   "ngPackage": {
